+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Release History
 ===============
 
+v1.2.1 - (2022-12-04)
+---------------------
+- Fix Python 2 incompatibility on the backend
+- Allow skipping prepare metadata for the full build by returning None as basename
+
 v1.2.0 - (2022-12-04)
 ---------------------
 - Expose which optional hooks are present or missing via :meth:`pyproject_api.Frontend.optional_hooks`